feat: Update dependencies and add bounceIn animation
Updates Firebase dependency to version 10.8.0 and introduces a new bounceIn animation to the UI. The animation is applied to the checkmark icon in the GoalsAchievementsScreen.

Additionally, the service worker cache version is incremented to v16 and a new cache for libraries (LIB_CACHE_NAME) is introduced to ensure proper cache management. The manifest link in index.html is also corrected.